+++ Reddit/Easy/making_links/src/lib.rs
@@ -1,16 +1,20 @@
 
 pub fn convert(input: &str) -> String {
-    let mut input = input.parse::<i32>().unwrap();
+    let letters: Vec<char> = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz"
+        .chars().collect();
+
+    let mut input = input.parse::<usize>().unwrap();
+
+    let mut result = String::new();
 
     while input > 0 {
-        println!("{}", input%62);
-
+        // println!("{} -> {}", input%62, letters[input%62]);
+        result.push(letters[input%62]);
         input /= 62;
     }
 
-    // 0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ
 
-    "".to_string()
+    result.chars().rev().collect()
 }
 
 #[cfg(test)]